refactor: change directory index
